server/requirements.txt
fastapi
uvicorn
websockets
python-multipart
opencv-python-headless
numpy
google-cloud-aiplatform
google-generativeai
google-cloud-texttospeech
pydantic
python-dotenv
aiofiles
ffmpeg-python
PyPDF2
python-docx
hypothesis
pytest
pytest-asyncio
